Got rendering working with no culling

+++ src/main/java/me/cortex/cullmister/RegionManager.java
@@ -1,6 +1,7 @@
 package me.cortex.cullmister;
 
 import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
+import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
 import me.cortex.cullmister.region.Region;
 import me.cortex.cullmister.region.RegionPos;
 import net.caffeinemc.sodium.render.chunk.compile.ChunkBuilder;
@@ -8,6 +9,7 @@ import net.caffeinemc.sodium.render.chunk.compile.tasks.TerrainBuildResult;
 import net.caffeinemc.sodium.render.chunk.compile.tasks.TerrainBuildTask;
 import net.caffeinemc.sodium.render.chunk.passes.ChunkRenderPassManager;
 import net.caffeinemc.sodium.render.terrain.format.TerrainVertexFormats;
+import net.caffeinemc.sodium.world.ChunkTracker;
 import net.caffeinemc.sodium.world.slice.WorldSliceData;
 import net.caffeinemc.sodium.world.slice.cloned.ClonedChunkSectionCache;
 import net.minecraft.client.world.ClientWorld;
@@ -19,43 +21,77 @@ public class RegionManager {
 
     public ChunkBuildEngine builder;
 
-    public RegionManager(ClientWorld world) {
-        builder = new ChunkBuildEngine(world);
-    }
 
+    ObjectOpenHashSet<ChunkSectionPos> chunkSectionsNonImportant = new ObjectOpenHashSet<>();
+    ObjectOpenHashSet<ChunkSectionPos> chunkSectionsImportant = new ObjectOpenHashSet<>();
     public void tick(int frame) {
+        if (builder == null) {
+            return;
+        }
+        chunkSectionsImportant.forEach(pos->builder.requestRebuild(pos, frame, true));
+        chunkSectionsNonImportant.forEach(pos->builder.requestRebuild(pos, frame, false));
+        chunkSectionsImportant.clear();
+        chunkSectionsNonImportant.clear();
+
         synchronized (builder.outflowWorkQueue) {
             while (!builder.outflowWorkQueue.isEmpty()) {
                 TerrainBuildResult result = builder.outflowWorkQueue.dequeue();
+                //TODO: FIX, THIS IS A HACK
+                if (result.geometry().vertices() == null) {
+                    result.delete();
+                    continue;
+                }
                 getRegion(result.pos()).updateMeshes(result);
                 result.delete();
             }
         }
     }
 
+    public void setWorld(ClientWorld world) {
+        if (world == null)
+            return;
+        if (builder != null) {
+            builder.delete();
+        }
+        builder = new ChunkBuildEngine(world);
+    }
+
+    public void reset() {
+        for (Region r : regions.values()) {
+            r.delete();
+        }
+        regions.clear();
+        if (builder != null) {
+            builder.clear();
+        }
+    }
 
     public Region getRegion(ChunkSectionPos pos) {
         return getRegion(RegionPos.from(pos));
     }
 
     public Region getRegion(RegionPos pos) {
-        return regions.computeIfAbsent(pos.Long(), p->new Region(pos));
+        Region r = regions.computeIfAbsent(pos.Long(), p->new Region(pos));
+        if (!r.pos.equals(pos))
+            throw new IllegalStateException();
+        return r;
     }
 
     public void enqueueRemoval(ChunkSectionPos pos) {
         if (!regions.containsKey(RegionPos.from(pos).Long()))
             throw new IllegalStateException();
-        //TODO: if region chunk count is zero, remove the region
+
         Region r = getRegion(pos);
         r.remove(pos);
         if (r.sectionCount == 0) {
-            System.out.println("Region deletion: "+ RegionPos.from(pos));
+            //System.out.println("Region deletion: "+ RegionPos.from(pos));
             regions.remove(RegionPos.from(pos).Long());
+            r.delete();
         }
     }
 
-    public void enqueueRebuild(ChunkSectionPos pos, int frame) {
-        builder.requestRebuild(pos, frame);
+    public void enqueueRebuild(ChunkSectionPos pos, boolean important) {
+        (important?chunkSectionsImportant:chunkSectionsNonImportant).add(pos);
     }
 
 }
